Operational Multithreaded Port Scanner
This port scanner automatically scans every single port on your own IP and prints out all ports that are open along with how much time in minutes it took to find that open port. Currently it takes just under 20 minutes to scan all 65k ports but I have an idea of how to fix that.